Harden order checkout after code review
Review findings applied: map the versioned cart-delete race to 409
ORDER_CONFLICT instead of a 500 through the global catch-all; widen the
order total to BIGINT/long so a large cart cannot silently overflow;
pin line-item ordering with @OrderBy; move the empty-cart invariant
into Order.place as the single source (it now throws EmptyCartException
directly); bind the snapshot product once in OrderItem.from.

+++ src/main/java/dev/zakalren/pickmeup/order/dto/OrderResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public record OrderResponse(
         Long id,
-        Integer totalPrice,
+        Long totalPrice,
         LocalDateTime orderedAt,
         List<OrderItemResponse> items
 ) {
